requirements.txt
Django-4.1.3
Python-3.10.8
Djangorestframework-3.14.0
Postgresql-14
psycopg2-binary-2.9.5
coverage-6.5.0
django-filter-22.1
Python Social Auth